HallThruster.Air

HallThruster.Argon

HallThruster.Electron

HallThruster.Krypton

HallThruster.NA

HallThruster.R0

HallThruster.Xenon

HallThruster.e

HallThruster.kB

HallThruster.mₑ

HallThruster.Gas

HallThruster.Gas

HallThruster.Species

HallThruster.B_field

HallThruster.ContinuityOnly

HallThruster.EulerEquations

HallThruster.IsothermalEuler

HallThruster.cf_electron_transport

HallThruster.channel_area

HallThruster.channel_area

HallThruster.electron_pressure

HallThruster.generate_grid

HallThruster.get_v_an

HallThruster.get_v_c

HallThruster.ln_λ

HallThruster.reconstruct!

HallThruster.solve_potential!

HallThruster.σ_en

HallThruster.Air — Constant

Air::Gas

Earth air at standard temperature and pressure

HallThruster.Argon — Constant

Argon::Gas

Argon gas

HallThruster.Electron — Constant

Electron::Species

Electron

HallThruster.Krypton — Constant

Krypton::Gas

Krypton gas

HallThruster.NA — Constant

NA

Number of atoms in a kg-mol (6.02214076e26 / kmol)

HallThruster.R0 — Constant

R0

Universal gas constant (8314.46261815324 J / kmol K)

HallThruster.Xenon — Constant

Xenon::Gas

Xenon gas

HallThruster.e — Constant

e

Electron charge (1.602176634e-19 Coulomb)

HallThruster.kB — Constant

kB

Boltzmann constant (1.380649e-23 J/K)

HallThruster.mₑ — Constant

mₑ

Electron mass (9.10938356e-31 kilograms)

HallThruster.Gas — Type

Gas

A chemical element in the gaseous state. Container for element properties used in fluid computations.

Fields

name::String Full name of gas (i.e. Xenon)

short_name::String Short name/symbol (i.e. Xe for Xenon)

γ::Float64 Specific heat ratio / adiabatic index

M::Float64 Molar mass (grams/mol) or atomic mass units

m::Float64 Mass of atom in kg

cp::Float64 Specific heat at constant pressure in J / kg / K

cv::Float64 Specific heat at constant volume in J / kg / K

R::Float64 Gas constant in J / kg / K

HallThruster.Gas — Method

Gas(name::String, short_name::String; γ::Float64, M::Float64)

Instantiate a new Gas, providing a name, short name, the adiabatic index, and the molar mass. Other gas properties, including gas constant, specific heats at constant pressure/volume, and mass of atom/molecule in kg will are then computed.

julia> Gas("Xenon", "Xe", γ = 5/3, M = 83.798)
Xenon

HallThruster.Species — Type

Species

Represents a gas with a specific charge state. In a plasma, different ionization states of the same gas may coexist, so we need to be able to differentiate between these.

julia> Species(Xenon, 0)
Xe

julia> Species(Xenon, 1)
Xe+

julia> Species(Xenon, 3)
Xe3+

HallThruster.B_field — Method

B_field(B_max::Float64, z::Float64, L_ch::Float64)

defines magnetic field as a function of position.

HallThruster.ContinuityOnly — Method

ContinuityOnly

A ConservationLawSystem in which only continuity (mass conservation) is solved, while velocity and temperature are held constant. Must specify a constant velocity (in m/s) and temperature (in K).

julia> equation = ContinuityOnly(u = 300, T = 500)
ContinuityOnly(u = 300.0 m/s, T = 500.0 K)

HallThruster.EulerEquations — Method

EulerEquations

A ConservationLawSystem for the inviscid Navier-Stokes equations, better known as the Euler equations. Velocity and temperature are variable, so the values held in the ConservationLawSystem are set to zero and subsequently unused.

julia> equation = EulerEquations()
EulerEquations()

HallThruster.IsothermalEuler — Method

IsothermalEuler

A ConservationLawSystem in which only continuity and inviscid momentum are solved, while temperature is held constant. Must specify a constant temperature (in K).

julia> equation = IsothermalEuler(T = 500)
IsothermalEuler(T = 500.0 K)

HallThruster.cf_electron_transport — Method

cf_electron_transport(v_an::Float64, v_c::Float64, B::Float64)

calculates electron transport according to the generalized Ohm's law as a function of the classical and anomalous collision frequencies and the magnetic field.

HallThruster.channel_area — Method

channel_area(outer_radius, inner_radius, length)

Compute the area of a Hall thruster channel from its dimensions

HallThruster.channel_area — Method

channel_area(geometry::Geometry1D)

Compute the area of the Hall thruster channel from the given Geometry1D object

HallThruster.electron_pressure — Method

electron_pressure(ne::Float64, Tev::Float64)

ideal gas law for electrons. Tev is in eV.

HallThruster.generate_grid — Method

generate_grid(geometry, ncells)

Generate a one-dimensional uniform grid on the domain specified in the geomety. Returns number of cells, coordinates of cell centers (plus ghost cells face coordinates), interface/edges and volume of a cell for number density calculations.

HallThruster.get_v_an — Method

get_v_an(z::Float64, B::Float64, L_ch::Float64)

defines model for anomalous collision frequency.

HallThruster.get_v_c — Method

get_v_c(Tev::Float64, nn::Float64, ne::Float64, m::Float64)

calculate classical collision frequency, consisting of electron neutral and electron ion collision frequencies. Eq. 3.6-12 and 3.6-14, from Fundamentals of Electric Propulsion, Goebel and Katz, 2008.

HallThruster.ln_λ — Method

ln_λ(ne::Float64, Tev::Float64)

calculate coulomb logarithm as a function of electron number density and electron temperature in eV. Eq. 3.6-15, from Fundamentals of Electric Propulsion, Goebel and Katz, 2008.

HallThruster.reconstruct! — Method

reconstruct!

Reconstruction using the MUSCL scheme. UL is the flux to the left, therefore evaluated on the right face, UR is the flux to the right, therefore evaluated on the left face.

HallThruster.solve_potential! — Method

solve_potential!(; ϕ::Vector{Float64}, U::Matrix{Float64}, params::NamedTuple)

function to solve the potential equation derived from the generalized Ohm's law and employing charge conservation using quasineutrality. Second derivatives approximated with 2nd order central difference scheme, first derivatives with central difference. Centers of computational mesh equivalent with centers of fluid mesh, but omitting boundary ghost cells. If required, interpolation is used to infer properties at mesh boundaries. Potential is a function of magnetic field, anomalous and classical collision frequency, neutral and ion density as well as ion velocity, and electron density and temperature leading to electron pressure. Solved by inverting a tridiagonal matrix.

HallThruster.σ_en — Method

σ_en(Tev::Float64)

calculation electron neutral collision cross section in m² as a function of electron temperature in eV. Eq. 3.6-13, from Fundamentals of Electric Propulsion, Goebel and Katz, 2008.