HallThruster.Air

HallThruster.Argon

HallThruster.Electron

HallThruster.Krypton

HallThruster.NA

HallThruster.R0

HallThruster.Xenon

HallThruster.e

HallThruster.kB

HallThruster.mₑ

HallThruster.EnergyOVS

HallThruster.Gas

HallThruster.Gas

HallThruster.Species

HallThruster.Verification

HallThruster.B_field

HallThruster.ContinuityOnly

HallThruster.EulerEquations

HallThruster.IsothermalEuler

HallThruster.backward_diff_coeffs

HallThruster.backward_difference

HallThruster.central_diff_coeffs

HallThruster.central_difference

HallThruster.channel_area

HallThruster.channel_area

HallThruster.coulomb_logarithm

HallThruster.downwind_diff_coeffs

HallThruster.electron_collision_freq

HallThruster.electron_mobility

HallThruster.electron_pressure

HallThruster.forward_diff_coeffs

HallThruster.forward_difference

HallThruster.freq_electron_electron

HallThruster.freq_electron_ion

HallThruster.freq_electron_neutral

HallThruster.generate_grid

HallThruster.interpolation_coeffs

HallThruster.lerp

HallThruster.read_restart

HallThruster.second_deriv_central_diff

HallThruster.second_deriv_coeffs

HallThruster.solve_potential_edge!

HallThruster.upwind_diff_coeffs

HallThruster.write_restart

HallThruster.σ_en

HallThruster.Air — Constant

Air::Gas

Earth air at standard temperature and pressure

HallThruster.Argon — Constant

Argon::Gas

Argon gas

HallThruster.Electron — Constant

Electron::Species

Electron

HallThruster.Krypton — Constant

Krypton::Gas

Krypton gas

HallThruster.NA — Constant

NA

Number of atoms in a kg-mol (6.02214076e26 / kmol)

HallThruster.R0 — Constant

R0

Universal gas constant (8314.46261815324 J / kmol K)

HallThruster.Xenon — Constant

Xenon::Gas

Xenon gas

HallThruster.e — Constant

e

Electron charge (1.602176634e-19 Coulomb)

HallThruster.kB — Constant

kB

Boltzmann constant (1.380649e-23 J/K)

HallThruster.mₑ — Constant

mₑ

Electron mass (9.10938356e-31 kilograms)

HallThruster.EnergyOVS — Type

mutable struct EnergyOVS

Enables setting mu, ue, Tev and ne to certain values to very electron energy equation

HallThruster.Gas — Type

Gas

A chemical element in the gaseous state. Container for element properties used in fluid computations.

Fields

name::String Full name of gas (i.e. Xenon)

short_name::String Short name/symbol (i.e. Xe for Xenon)

γ::Float64 Specific heat ratio / adiabatic index

M::Float64 Molar mass (grams/mol) or atomic mass units

m::Float64 Mass of atom in kg

cp::Float64 Specific heat at constant pressure in J / kg / K

cv::Float64 Specific heat at constant volume in J / kg / K

R::Float64 Gas constant in J / kg / K

HallThruster.Gas — Method

Gas(name::String, short_name::String; γ::Float64, M::Float64)

Instantiate a new Gas, providing a name, short name, the adiabatic index, and the molar mass. Other gas properties, including gas constant, specific heats at constant pressure/volume, and mass of atom/molecule in kg will are then computed.

julia> Gas("Xenon", "Xe", γ = 5/3, M = 83.798)
Xenon

HallThruster.Species — Type

Species

Represents a gas with a specific charge state. In a plasma, different ionization states of the same gas may coexist, so we need to be able to differentiate between these.

julia> Species(Xenon, 0)
Xe

julia> Species(Xenon, 1)
Xe+

julia> Species(Xenon, 3)
Xe3+

HallThruster.Verification — Type

mutable struct Verification

is passed to params to identify if OVS is active.

HallThruster.B_field — Method

B_field(B_max::Float64, z::Float64, L_ch::Float64)

defines magnetic field as a function of position.

HallThruster.ContinuityOnly — Method

ContinuityOnly

A ConservationLawSystem in which only continuity (mass conservation) is solved, while velocity and temperature are held constant. Must specify a constant velocity (in m/s) and temperature (in K).

julia> equation = ContinuityOnly(u = 300, T = 500)
_ContinuityOnly(u = 300.0 m/s, T = 500.0 K)

HallThruster.EulerEquations — Method

EulerEquations

A ConservationLawSystem for the inviscid Navier-Stokes equations, better known as the Euler equations. Velocity and temperature are variable, so the values held in the ConservationLawSystem are set to zero and subsequently unused.

julia> equation = EulerEquations()
_EulerEquations()

HallThruster.IsothermalEuler — Method

IsothermalEuler

A ConservationLawSystem in which only continuity and inviscid momentum are solved, while temperature is held constant. Must specify a constant temperature (in K).

julia> equation = IsothermalEuler(T = 500)
_IsothermalEuler(T = 500.0 K)

HallThruster.backward_diff_coeffs — Method

backward_diff_coeffs(x0, x1, x2)

Generate finite difference coefficients for a backward first derivative approximation at the point x2 on a three-point stencil at points x0, x1, and x2

julia> backward_diff_coeffs(-2//1, -1//1, 0//1)
(1//2, -2//1, 3//2)
julia> backward_diff_coeffs(-3//2, -1//1, 0//1)
(4//3, -3//1, 5//3)

HallThruster.backward_difference — Method

backward_difference(f0, f1, f2, x0, x1, x2)

Given three points x0, x1, and x2, and the function values at those points, f0, f1, f2, compute the second-order approximation of the derivative at x2

f(x) = x^4
x0, x1, x2 = 1.9999998, 1.9999999, 2
bd = backward_difference(f(x0), f(x1), f(x2), x0, x1, x2)
bd ≈ 32

# output

true

HallThruster.central_diff_coeffs — Method

central_diff_coeffs(x0, x1, x2)

Generate finite difference coefficients for a central first derivative approximation at the point x1 on a three-point stencil at points x0, x1, and x2

julia> central_diff_coeffs(-1//1, 0//1, 1//1)
(-1//2, 0//1, 1//2)
julia> central_diff_coeffs(-1//2, 0//1, 1//1)
(-4//3, 1//1, 1//3)

HallThruster.central_difference — Method

central_difference(f0, f1, f2, x0, x1, x2)

Given three points x0, x1, and x2, and the function values at those points, f0, f1, f2, compute the second-order approximation of the derivative at x1

f(x) = x^4
x0, x1, x2 = 1.9999999, 2, 2.0000001
cd = central_difference(f(x0), f(x1), f(x2), x0, x1, x2)
cd ≈ 32

# output

true

HallThruster.channel_area — Method

channel_area(outer_radius, inner_radius, length)

Compute the area of a Hall thruster channel from its dimensions

HallThruster.channel_area — Method

channel_area(geometry::Geometry1D)

Compute the area of the Hall thruster channel from the given Geometry1D object

HallThruster.coulomb_logarithm — Function

coulomb_logarithm(ne, Tev, Z = 1)

calculate coulomb logarithm for electron-ion collisions as a function of ion charge state Z, electron number density in m^-3, and electron temperature in eV.

HallThruster.downwind_diff_coeffs — Method

downwind_diff_coeffs(x0, x1, x2)

Generate finite difference coefficients for a downwind first derivative approximation at the point x2 on a three-point stencil at points x0, x1, and x2 (uses only points x1 and x2)

julia> downwind_diff_coeffs(-1//1, 0//1, 2//1)
(0//1, -1//2, 1//2)

HallThruster.electron_collision_freq — Method

electron_collision_freq(Tev::Float64, nn::Float64, ne::Float64, m::Float64)

calculate classical collision frequency, consisting of electron neutral and electron ion collision frequencies. Eq. 3.6-12 and 3.6-14, from Fundamentals of Electric Propulsion, Goebel and Katz, 2008.

HallThruster.electron_mobility — Method

cf_electron_transport(v_an::Float64, v_c::Float64, B::Float64)

calculates electron transport according to the generalized Ohm's law as a function of the classical and anomalous collision frequencies and the magnetic field.

HallThruster.electron_pressure — Method

electron_pressure(ne::Float64, Tev::Float64)

ideal gas law for electrons. Tev is in eV.

HallThruster.forward_diff_coeffs — Method

forward_diff_coeffs(x0, x1, x2)

Generate finite difference coefficients for a forward first derivative approximation at the point x0 on a three-point stencil at points x0, x1, and x2

julia> forward_diff_coeffs(1.0, 2.0, 3.0)
(-1.5, 2.0, -0.5)
julia> forward_diff_coeffs(0//1, 1//2, 3//2)
(-8//3, 3//1, -1//3)

HallThruster.forward_difference — Method

forward_difference(f0, f1, f2, x0, x1, x2)

Given three points x0, x1, and x2, and the function values at those points, f0, f1, f2, compute the second-order approximation of the derivative at x0

f(x) = x^4
x0, x1, x2 = 2.0, 2.000001, 2.000002
fd = forward_difference(f(x0), f(x1), f(x2), x0, x1, x2)
fd ≈ 32

# output

true

HallThruster.freq_electron_electron — Method

freq_electron_electron(ne, Tev)

Effective frequency at which electrons are scattered due to collisions with other electrons

HallThruster.freq_electron_ion — Method

freq_electron_ion(ne, Tev, Z)

Effective frequency at which electrons are scattered due to collisions with ions

HallThruster.freq_electron_neutral — Method

freq_electron_neutral(nn, Tev)

Effective frequency of electron scattering caused by collisions with neutrals

HallThruster.generate_grid — Function

generate_grid(geometry, ncells)

Generate a one-dimensional uniform grid on the domain specified in the geomety. Returns number of cells, coordinates of cell centers (plus ghost cells face coordinates), interface/edges and volume of a cell for number density calculations.

HallThruster.interpolation_coeffs — Method

interpolation_coeffs(x, x0, x1, y0, y1)

Compute the coefficients for interpolation between two points (x0, y0) and (x1, y1) such that y = c0 * y0 + c1 * y1 ```jldoctest;setup = :(using HallThruster: itpcoeffs, lerp) julia> c0, c1 = interpolationcoeffs(0.5, 0.0, 1.0, 0.0, 2.0) (0.5, 0.5) julia> c0 * 0.0 + c1 * 2.0 == lerp(0.5, 0.0, 1.0, 0.0, 2.0) true

HallThruster.lerp — Method

lerp(x, x0, x1, y0, y1)

Interpolate between two points (x0, y0) and (x1, y1) ```jldoctest;setup = :(using HallThruster: lerp) julia> lerp(0.5, 0.0, 1.0, 0.0, 2.0) 1.0

HallThruster.read_restart — Method

read_restart(path::AbstractString)

Load a restart file from path.

The filetype can be anything supported by FileIO, though JLD2 is preferred.

HallThruster.second_deriv_central_diff — Method

second_deriv_central_diff(f0, f1, f2, x0, x1, x2)

Given three points x0, x1, and x2, and the function values at those points, f0, f1, f2, compute the second-order approximation of the second derivative at x1

f(x) = x^4
x0, x1, x2 = 1.9999, 2.0, 2.0001
sd = second_deriv_central_diff(f(x0), f(x1), f(x2), x0, x1, x2)
sd ≈ 48

# output

true

HallThruster.second_deriv_coeffs — Method

second_deriv_coeffs(x0, x1, x2)

Generate finite difference coefficients for a central second derivative approximation at the point x1 on a three-point stencil at points x0, x1, and x2

julia> second_deriv_coeffs(-2//1, 0//1, 2//1)
(1//4, -1//2, 1//4)
julia> second_deriv_coeffs(-1//2, 0//1, 1//1)
(8//3, -4//1, 4//3)

HallThruster.solve_potential_edge! — Method

solve_potential!(; U::Matrix{Float64}, params::NamedTuple)

function to solve the potential equation derived from the generalized Ohm's law and employing charge conservation using quasineutrality. Second derivatives approximated with 2nd order central difference scheme, first derivatives with central difference. Centers of computational mesh placed on edges of fluid mesh, therefore edges correspond to boundaries for fluid. If required, interpolation is used to infer properties at mesh boundaries. Potential is a function of magnetic field, anomalous and classical collision frequency, neutral and ion density as well as ion velocity, and electron density and temperature leading to electron pressure. Solved by applying Thomas algorithm, which is of complexity O(n) and valid if matrix tridiagonal and diagonally dominant. The latter assumption almost always holds unless there is a huge discontinuity in either electron mobility or electron density.

HallThruster.upwind_diff_coeffs — Method

upwind_diff_coeffs(x0, x1, x2)

Generate finite difference coefficients for a upwind first derivative approximation at the point x2 on a three-point stencil at points x0, x1, and x2 (uses only points x0 and x1)

julia> upwind_diff_coeffs(-3//1, 0//1, 2//1)
(-1//3, 1//3, 0//1)

HallThruster.write_restart — Method

write_restart(path::AbstractString, sol)

Write a restart file to path`.

This can be reloaded to resume a simulation. The filetype can be anything supported by FileIO, though JLD2 is preferred.

HallThruster.σ_en — Method

σ_en(Tev)

Electron neutral collision cross section in m² as a function of electron temperature in eV. Eq. 3.6-13, from Fundamentals of Electric Propulsion, Goebel and Katz, 2008.