HallThruster.Air

HallThruster.Argon

HallThruster.Electron

HallThruster.Krypton

HallThruster.NA

HallThruster.R0

HallThruster.Xenon

HallThruster.e

HallThruster.kB

HallThruster.Gas

HallThruster.Gas

HallThruster.Species

HallThruster.ContinuityOnly

HallThruster.EulerEquations

HallThruster.IsothermalEuler

HallThruster.channel_area

HallThruster.channel_area

HallThruster.reconstruct!

HallThruster.Air — Constant

Air::Gas

Earth air at standard temperature and pressure

HallThruster.Argon — Constant

Argon::Gas

Argon gas

HallThruster.Electron — Constant

Electron::Species

Electron

HallThruster.Krypton — Constant

Krypton::Gas

Krypton gas

HallThruster.NA — Constant

NA

Number of atoms in a kg-mol (6.02214076e26 / kmol)

HallThruster.R0 — Constant

R0

Universal gas constant (8314.46261815324 J / kmol K)

HallThruster.Xenon — Constant

Xenon::Gas

Xenon gas

HallThruster.e — Constant

e

Electron charge (1.602176634e-19 Coulomb)

HallThruster.kB — Constant

kB

Boltzmann constant (1.380649e-23 J/K)

HallThruster.Gas — Type

Gas

A chemical element in the gaseous state. Container for element properties used in fluid computations.

Fields

name::String Full name of gas (i.e. Xenon)

short_name::String Short name/symbol (i.e. Xe for Xenon)

γ::Float64 Specific heat ratio / adiabatic index

M::Float64 Molar mass (grams/mol) or atomic mass units

m::Float64 Mass of atom in kg

cp::Float64 Specific heat at constant pressure in J / kg / K

cv::Float64 Specific heat at constant volume in J / kg / K

R::Float64 Gas constant in J / kg / K

HallThruster.Gas — Method

Gas(name::String, short_name::String; γ::Float64, M::Float64)

Instantiate a new Gas, providing a name, short name, the adiabatic index, and the molar mass. Other gas properties, including gas constant, specific heats at constant pressure/volume, and mass of atom/molecule in kg will are then computed.

julia> Gas("Xenon", "Xe", γ = 5/3, M = 83.798)
Xenon

HallThruster.Species — Type

Species

Represents a gas with a specific charge state. In a plasma, different ionization states of the same gas may coexist, so we need to be able to differentiate between these.

julia> Species(Xenon, 0)
Xe

julia> Species(Xenon, 1)
Xe+

julia> Species(Xenon, 3)
Xe3+

HallThruster.ContinuityOnly — Method

ContinuityOnly

A ConservationLawSystem in which only continuity (mass conservation) is solved, while velocity and temperature are held constant. Must specify a constant velocity (in m/s) and temperature (in K).

julia> equation = ContinuityOnly(u = 300, T = 500)
ContinuityOnly(u = 300.0 m/s, T = 500.0 K)

HallThruster.EulerEquations — Method

EulerEquations

A ConservationLawSystem for the inviscid Navier-Stokes equations, better known as the Euler equations. Velocity and temperature are variable, so the values held in the ConservationLawSystem are set to zero and subsequently unused.

julia> equation = EulerEquations()
EulerEquations()

HallThruster.IsothermalEuler — Method

IsothermalEuler

A ConservationLawSystem in which only continuity and inviscid momentum are solved, while temperature is held constant. Must specify a constant temperature (in K).

julia> equation = IsothermalEuler(T = 500)
IsothermalEuler(T = 500.0 K)

HallThruster.channel_area — Method

channel_area(outer_radius, inner_radius, length)

Compute the area of a Hall thruster channel from its dimensions

HallThruster.channel_area — Method

channel_area(geometry::Geometry1D)

Compute the area of the Hall thruster channel from the given Geometry1D object

HallThruster.reconstruct! — Method

reconstruct!

Reconstruction using the MUSCL scheme. UL is the flux to the left, therefore evaluated on the right face, UR is the flux to the right, therefore evaluated on the left face.