HallThruster.Air

HallThruster.Argon

HallThruster.Bismuth

HallThruster.Electron

HallThruster.Krypton

HallThruster.Mercury

HallThruster.NA

HallThruster.R0

HallThruster.Xenon

HallThruster.e

HallThruster.kB

HallThruster.me

HallThruster.AnomalousTransportModel

HallThruster.Bohm

HallThruster.Config

HallThruster.ExcitationLookup

HallThruster.FixedAnomModel

HallThruster.Gas

HallThruster.Gas

HallThruster.Geometry1D

HallThruster.IonizationLookup

HallThruster.LandmarkExcitationLookup

HallThruster.LandmarkIonizationLookup

HallThruster.MultiLogBohm

HallThruster.NoAnom

HallThruster.NoExcitation

HallThruster.Species

HallThruster.Thruster

HallThruster.TwoZoneBohm

HallThruster.WallLossModel

HallThruster.WallMaterial

HallThruster.ZeroEquationModel

HallThruster.ContinuityOnly

HallThruster.EulerEquations

HallThruster.IsothermalEuler

HallThruster.SEE_yield

HallThruster.backward_diff_coeffs

HallThruster.backward_difference

HallThruster.central_diff_coeffs

HallThruster.central_difference

HallThruster.channel_area

HallThruster.channel_perimeter

HallThruster.channel_width

HallThruster.compute_current

HallThruster.coulomb_logarithm

HallThruster.downwind_diff_coeffs

HallThruster.electron_mobility

HallThruster.evaluate_anom

HallThruster.evaluate_anom

HallThruster.forward_diff_coeffs

HallThruster.forward_difference

HallThruster.freq_electron_electron

HallThruster.freq_electron_ion

HallThruster.freq_electron_neutral

HallThruster.generate_grid

HallThruster.initialize_anom!

HallThruster.initialize_anom!

HallThruster.interpolation_coeffs

HallThruster.lerp

HallThruster.load_reactions

HallThruster.load_reactions

HallThruster.maximum_charge_state

HallThruster.read_restart

HallThruster.run_simulation

HallThruster.second_deriv_central_diff

HallThruster.second_deriv_coeffs

HallThruster.sheath_potential

HallThruster.solve_potential_edge!

HallThruster.supported_gases

HallThruster.time_average

HallThruster.upwind_diff_coeffs

HallThruster.write_restart

HallThruster.σ_en

HallThruster.Air — Constant

Air::Gas

Earth air at standard temperature and pressure

HallThruster.Argon — Constant

Argon::Gas

Argon gas

HallThruster.Bismuth — Constant

Bismuth::Gas

Bismuth vapor

HallThruster.Electron — Constant

Electron::Species

Electron

HallThruster.Krypton — Constant

Krypton::Gas

Krypton gas

HallThruster.Mercury — Constant

Mercury::Gas

Mercury vapor

HallThruster.NA — Constant

NA

Number of atoms in a kg-mol (6.02214076e26 / kmol)

HallThruster.R0 — Constant

R0

Universal gas constant (8314.46261815324 J / kmol K)

HallThruster.Xenon — Constant

Xenon::Gas

Xenon gas

HallThruster.e — Constant

e

Electron charge (1.602176634e-19 Coulomb)

HallThruster.kB — Constant

kB

Boltzmann constant (1.380649e-23 J/K)

HallThruster.me — Constant

me

Electron mass (9.10938356e-31 kilograms)

HallThruster.AnomalousTransportModel — Type

AnomalousTransportModel

The abstract supertype of all types of anomalous transport model. Currently, there are two subtypes: FixedAnomModel and ZeroEquationModel, which correspond, respectively, to models which are initialized and then never updated, and models which depend algebraically on local plasma properties.

HallThruster.Bohm — Type

Bohm(c) <: FixedAnomModel

Model where the anomalous collision frequency scales with the electron cyclotron frequency ωce times some scaling factor c

HallThruster.Config — Type

struct Config{A<:HallThruster.AnomalousTransportModel, W<:HallThruster.WallLossModel, IZ<:HallThruster.IonizationModel, EX<:HallThruster.ExcitationModel, EN<:HallThruster.ElectronNeutralModel, HET<:HallThruster.Thruster, S_N, S_IC, S_IM, S_ϕ, S_E, T<:HallThruster.TransitionFunction, IC<:HallThruster.InitialCondition, CB, HS<:HallThruster.HyperbolicScheme}

Hall thruster configuration struct. Only four mandatory fields: discharge_voltage, thruster, anode_mass_flow_rate, and domain.

Fields

discharge_voltage::Float64

cathode_potential::Float64

anode_Te::Float64

cathode_Te::Float64

wall_loss_model::HallThruster.WallLossModel

neutral_velocity::Float64

neutral_temperature::Float64

implicit_energy::Float64

propellant::HallThruster.Gas

ncharge::Int64

ion_temperature::Float64

anom_model::HallThruster.AnomalousTransportModel

ionization_model::HallThruster.IonizationModel

excitation_model::HallThruster.ExcitationModel

electron_neutral_model::HallThruster.ElectronNeutralModel

electron_ion_collisions::Bool

electron_pressure_coupled::Float64

min_number_density::Float64

min_electron_temperature::Float64

transition_function::HallThruster.TransitionFunction

initial_condition::HallThruster.InitialCondition

callback::Any

magnetic_field_scale::Float64

source_neutrals::Any

source_ion_continuity::Any

source_ion_momentum::Any

source_potential::Any

source_energy::Any

scheme::HallThruster.HyperbolicScheme

thruster::HallThruster.Thruster

domain::Tuple{Float64, Float64}

LANDMARK::Bool

anode_mass_flow_rate::Float64

ion_wall_losses::Bool

HallThruster.ExcitationLookup — Type

ExcitationLookup(;[directories::Vector{String} = String[]])

Default excitation model for HallThruster.jl. Reads excitation rate coefficients from file. Looks (preferentially) in provided directories and in the reactions subfolder for rate coefficient files

HallThruster.FixedAnomModel — Type

FixedAnomModel <: AnomalousTransportModel

An anomalous collision frequency model which is fixed at the start of a simulation and then never updated again. This is the most common type of model, and includes the ubiquitous multi-zone bohm-like collision frequency models.

Users defining their own FixedAnomModel must implement the initialize_anom!(νan, model, U, params) method for their type.

HallThruster.Gas — Type

Gas

A chemical element in the gaseous state. Container for element properties used in fluid computations.

Fields

name::String Full name of gas (i.e. Xenon)

short_name::String Short name/symbol (i.e. Xe for Xenon)

γ::Float64 Specific heat ratio / adiabatic index

M::Float64 Molar mass (grams/mol) or atomic mass units

m::Float64 Mass of atom in kg

cp::Float64 Specific heat at constant pressure in J / kg / K

cv::Float64 Specific heat at constant volume in J / kg / K

R::Float64 Gas constant in J / kg / K

HallThruster.Gas — Method

Gas(name::String, short_name::String; γ::Float64, M::Float64)

Instantiate a new Gas, providing a name, short name, the adiabatic index, and the molar mass. Other gas properties, including gas constant, specific heats at constant pressure/volume, and mass of atom/molecule in kg will are then computed.

julia> Gas("Xenon", "Xe", γ = 5/3, M = 83.798)
Xenon

HallThruster.Geometry1D — Type

Geometry1D

Struct containing information about Hall thruster geometry. Required fields are channel_length, inner_radius, and outer_radius, all in meters. Contains a fourth field, channel_area, which is computed from the above three.

Fields

channel_length::Float64

inner_radius::Float64

outer_radius::Float64

channel_area::Float64

HallThruster.IonizationLookup — Type

IonizationLookup(;[directories::Vector{String} = String[]])

Default ionization model for HallThruster.jl. Reads ionization rate coefficients from file. Looks (preferentially) in provided directories and in the reactions subfolder for rate coefficient files

HallThruster.LandmarkExcitationLookup — Type

LandmarkExcitationLookup()

Excitation model for the LANDMARK benchmark.

Reads excitation rate coefficients from the landmark/landmark_rates.csv file in the HallThruster.jl main directory. Supports only singly-charged Xenon.

HallThruster.LandmarkIonizationLookup — Type

LandmarkIonizationLookup()

Ionization model for the LANDMARK benchmark.

Reads ionization rate coefficients from the landmark/landmark_rates.csv file in the HallThruster.jl main directory. Supports only singly-charged Xenon.

HallThruster.MultiLogBohm — Type

MultiLogBohm(zs, cs) <: FixedAnomModel

Model similar to that employed in Hall2De, where the mobility is Bohm-like (i.e. νan(z) = c(z) * ωce(z)) and z is in meters.

The function c(z) is defined by a sequence of nodes (z, c) provided by the user. At z = z[1], c(z) = c[1], and so forth.

At z[i] < z < z[i+1], log(c) is defined by linearly interpolating between log(c[i]) and log(c[i+1]).

For z < z[1], c = c[1] and for z > z[end], c(z) = c[end].

The user may also provide a single array of [z[1], z[2], ..., z[end], c[1], c[2], ..., c[end]]. The number of z values must be equal to the number of c values.

HallThruster.NoAnom — Type

NoAnom <: FixedAnomModel

No anomalous collision frequency included in simulation

HallThruster.NoExcitation — Type

NoExcitation <: ExcitationModel

Model for neglecting excitation energy losses

HallThruster.Species — Type

Species

Represents a gas with a specific charge state. In a plasma, different ionization states of the same gas may coexist, so we need to be able to differentiate between these.

julia> Species(Xenon, 0)
Xe

julia> Species(Xenon, 1)
Xe+

julia> Species(Xenon, 3)
Xe3+

HallThruster.Thruster — Type

Thruster

Struct containing information about a Hall thruster. This includes a name, geometry (a Geometry1D object), magnetic_field (radial magnetic field along centerline, a function which takes z in meters and outputs B in Tesla), and a shielded (a flag indicating whether the thruster is magnetically-shielded).

Fields

name::String

geometry::HallThruster.Geometry1D

magnetic_field::Any

shielded::Bool

HallThruster.TwoZoneBohm — Type

TwoZoneBohm(c1, c2) <: FixedAnomModel

Model where the anomalous collision frequency has two values: c1 * ωce inside the channel and c2 * ωce outside of the channel. Takes two arguments: c1 and c2. The transition between these values can be smoothed by the user-provided transition function.

HallThruster.WallLossModel — Type

abstract type WallLossModel

Abstract type for wall loss models in the electron energy equation. Types included with HallThruster.jl are:

NoWallLosses No electron energy losses to the wall.

ConstantSheathPotential(sheath_potential, inner_loss_coeff, outer_loss_coeff) Power to the walls scales with α * exp(-sheath_potential)), where α = inner_loss_coeff inside the channel and α = outer_loss_coeff outside.

WallSheath(material::WallMaterial) Power to the walls is computed self-consistently using a sheath model, dependent on the secondary electron emission yield of the provided material. See WallMaterial for material options.

Users implementing their own WallLossModel will need to implement at least three methods 1) freq_electron_wall(model, U, params, i): Compute the electron-wall momentum transfer collision frequency in cell i 2) wall_electron_current(model, U, params, i: Compute the electron current to the walls in cell i 2) wall_power_loss(model, U, params, i): Compute the electron power lost to the walls

A fourth method, wall_ion_current(model, U, params, i, Z), for computing the current of ions of charge Z to the walls in cell i, may also be implemented. If left unimplemented, it will default to computing the current assuming Ie,w = Ii,w.

HallThruster.WallMaterial — Type

struct WallMaterial

Struct containing secondary electron emission (SEE) yield fit coefficients for a material. Used in the WallSheath wall loss model

Fields

a::Float64

b::Float64

Γ::Float64

The SEE yield is computed as a function of electron temperature in eV (Tev) as Γ * a * Tev^b.

Available materials

IdealDielectric (a material with zero SEE)

Alumina

BoronNitride

BNSiO2

StainlessSteel

HallThruster.ZeroEquationModel — Type

ZeroEquationModel <: AnomalousTransportModel

An anomalous collision frequency which depends algebraically on the local plasma properties and is updated at each iteration.

Users defining their own ZeroEquationModel must implement the evaluate_anom(model, U, params) method for their type.

HallThruster.ContinuityOnly — Method

ContinuityOnly

A ConservationLawSystem in which only continuity (mass conservation) is solved, while velocity and temperature are held constant. Must specify a constant velocity (in m/s) and temperature (in K).

julia> equation = ContinuityOnly(u = 300, T = 500)
_ContinuityOnly(u = 300.0 m/s, T = 500.0 K)

HallThruster.EulerEquations — Method

EulerEquations

A ConservationLawSystem for the inviscid Navier-Stokes equations, better known as the Euler equations. Velocity and temperature are variable, so the values held in the ConservationLawSystem are set to zero and subsequently unused.

julia> equation = EulerEquations()
_EulerEquations()

HallThruster.IsothermalEuler — Method

IsothermalEuler

A ConservationLawSystem in which only continuity and inviscid momentum are solved, while temperature is held constant. Must specify a constant temperature (in K).

julia> equation = IsothermalEuler(T = 500)
_IsothermalEuler(T = 500.0 K)

HallThruster.SEE_yield — Method

SEE_yield(material::WallMaterial, Tev, mi)

fit function for SEE with different wall materials Goebel Katz equ. 7.3-30

HallThruster.backward_diff_coeffs — Method

backward_diff_coeffs(x0, x1, x2)

Generate finite difference coefficients for a backward first derivative approximation at the point x2 on a three-point stencil at points x0, x1, and x2

julia> backward_diff_coeffs(-2//1, -1//1, 0//1)
(1//2, -2//1, 3//2)
julia> backward_diff_coeffs(-3//2, -1//1, 0//1)
(4//3, -3//1, 5//3)

HallThruster.backward_difference — Method

backward_difference(f0, f1, f2, x0, x1, x2)

Given three points x0, x1, and x2, and the function values at those points, f0, f1, f2, compute the second-order approximation of the derivative at x2

f(x) = x^4
x0, x1, x2 = 1.9999998, 1.9999999, 2
bd = backward_difference(f(x0), f(x1), f(x2), x0, x1, x2)
bd ≈ 32

# output

true

HallThruster.central_diff_coeffs — Method

central_diff_coeffs(x0, x1, x2)

Generate finite difference coefficients for a central first derivative approximation at the point x1 on a three-point stencil at points x0, x1, and x2

julia> central_diff_coeffs(-1//1, 0//1, 1//1)
(-1//2, 0//1, 1//2)
julia> central_diff_coeffs(-1//2, 0//1, 1//1)
(-4//3, 1//1, 1//3)

HallThruster.central_difference — Method

central_difference(f0, f1, f2, x0, x1, x2)

Given three points x0, x1, and x2, and the function values at those points, f0, f1, f2, compute the second-order approximation of the derivative at x1

f(x) = x^4
x0, x1, x2 = 1.9999999, 2, 2.0000001
cd = central_difference(f(x0), f(x1), f(x2), x0, x1, x2)
cd ≈ 32

# output

true

HallThruster.channel_area — Method

channel_area(outer_radius, inner_radius)

Compute the cross-sectional area of a Hall thruster channel from its dimensions

HallThruster.channel_perimeter — Method

channel_perimeter(outer_radius, inner_radius)

Compute the perimeteter of the thruster channel, equal to the sum of the inner and outer circumferences

HallThruster.channel_width — Method

channel_width(outer_radius, inner_radius)

Compute the thruster channel width

HallThruster.compute_current — Function

compute_current(sol, location)

compute current at anode or cathode = outflow in 1D code.

HallThruster.coulomb_logarithm — Function

coulomb_logarithm(ne, Tev, Z = 1)

calculate coulomb logarithm for electron-ion collisions as a function of ion charge state Z, electron number density in m^-3, and electron temperature in eV.

HallThruster.downwind_diff_coeffs — Method

downwind_diff_coeffs(x0, x1, x2)

Generate finite difference coefficients for a downwind first derivative approximation at the point x2 on a three-point stencil at points x0, x1, and x2 (uses only points x1 and x2)

julia> downwind_diff_coeffs(-1//1, 0//1, 2//1)
(0//1, -1//2, 1//2)

HallThruster.electron_mobility — Method

electron_mobility(νan::Float64, νc::Float64, B::Float64)

calculates electron transport according to the generalized Ohm's law as a function of the classical and anomalous collision frequencies and the magnetic field.

HallThruster.evaluate_anom — Method

evaluate_anom(model::FixedAnomModel, U, params, i)

By default, this does nothing for FixedAnomModel objects, as the anomalous collision frequency is not updated at each iteration.

HallThruster.evaluate_anom — Method

evaluate_anom(model::ZeroEquationModel, U, params, i)

Return the anomalous collision frequency at index i for a provided ZeroEquationModel. Must be implemented by the user.

HallThruster.forward_diff_coeffs — Method

forward_diff_coeffs(x0, x1, x2)

Generate finite difference coefficients for a forward first derivative approximation at the point x0 on a three-point stencil at points x0, x1, and x2

julia> forward_diff_coeffs(1.0, 2.0, 3.0)
(-1.5, 2.0, -0.5)
julia> forward_diff_coeffs(0//1, 1//2, 3//2)
(-8//3, 3//1, -1//3)

HallThruster.forward_difference — Method

forward_difference(f0, f1, f2, x0, x1, x2)

Given three points x0, x1, and x2, and the function values at those points, f0, f1, f2, compute the second-order approximation of the derivative at x0

f(x) = x^4
x0, x1, x2 = 2.0, 2.000001, 2.000002
fd = forward_difference(f(x0), f(x1), f(x2), x0, x1, x2)
fd ≈ 32

# output

true

HallThruster.freq_electron_electron — Method

freq_electron_electron(ne, Tev)

Effective frequency at which electrons are scattered due to collisions with other electrons

HallThruster.freq_electron_ion — Method

freq_electron_ion(ne, Tev, Z)

Effective frequency at which electrons are scattered due to collisions with ions

HallThruster.freq_electron_neutral — Method

freq_electron_neutral(model::ElectronNeutralModel, nn, Tev)

Effective frequency of electron scattering caused by collisions with neutrals

HallThruster.generate_grid — Method

generate_grid(geometry, ncells)

Generate a one-dimensional uniform grid on the domain specified in the geomety. Returns number of cells, coordinates of cell centers (plus ghost cells face coordinates), interface/edges and volume of a cell for number density calculations.

HallThruster.initialize_anom! — Method

initialize_anom!(νan, model::FixedAnomModel, U, params)

Initialize the anomalous collision frequency profile for models in which the anomalous mobility does not change with time. Must be implemented by the user for specific subtypes of FixedAnomModel. Should update argument νan in-place with the desired value of anomalous collision frequency and then return the mutated array.

HallThruster.initialize_anom! — Method

initialize_anom!(νan, model::ZeroEquationModel, U, params)

Default anomalous collision frequency initialization for ZeroEquationModels. Just calls evalute_anom at each grid location.

HallThruster.interpolation_coeffs — Method

interpolation_coeffs(x, x0, x1, y0, y1)

Compute the coefficients for interpolation between two points (x0, y0) and (x1, y1) such that y = c0 * y0 + c1 * y1 ```jldoctest;setup = :(using HallThruster: itpcoeffs, lerp) julia> c0, c1 = interpolationcoeffs(0.5, 0.0, 1.0, 0.0, 2.0) (0.5, 0.5) julia> c0 * 0.0 + c1 * 2.0 == lerp(0.5, 0.0, 1.0, 0.0, 2.0) true

HallThruster.lerp — Method

lerp(x, x0, x1, y0, y1)

Interpolate between two points (x0, y0) and (x1, y1) ```jldoctest;setup = :(using HallThruster: lerp) julia> lerp(0.5, 0.0, 1.0, 0.0, 2.0) 1.0

HallThruster.load_reactions — Method

Charge exchange fits from Hause, Prince and Bemish, 2013

HallThruster.load_reactions — Method

load_reactions(model::ReactionModel, species)::Vector{IonizationReaction}

Load ionization reactions for the provided species and ionization model

HallThruster.maximum_charge_state — Method

maximum_charge_state(model::ReactionModel)::Int

Return the maximum supported charge state for a given reaction model. If 0 is returned, then no charge state restriction is applied.

HallThruster.read_restart — Method

read_restart(path::AbstractString)

Load a restart file from path.

The filetype can be anything supported by FileIO, though JLD2 is preferred.

HallThruster.run_simulation — Method

run_simulation(config; dt, duration, ncells, nsave, alg, restart, CFL, adaptive)

Run a Hall thruster simulation using the provided Config object.

Arguments

config: a Config containing simulation parameters.

dt: The timestep, in seconds. Typical values are O(10 ns) (1e-8 seconds).

duration: How long to run the simulation, in seconds (simulation time, not wall time). Typical runtimes are O(1 ms) (1e-3 seconds).

ncells: How many cells to use. Typical values are 100 - 1000 cells.

nsave: How many frames to save.

alg: Explicit timestepping algorithm to use. Defaults to second-order strong-stability preserving Runge-Kutta with a stage limiter and a step limiter(SSPRK22(stage_limiter = stage_limiter!)), but will work with any explicit timestepping algorithm provided by OrdinaryDiffEq.

restart: path to restart file or a HallThrusterSolution object. Defaults to nothing.

HallThruster.second_deriv_central_diff — Method

second_deriv_central_diff(f0, f1, f2, x0, x1, x2)

Given three points x0, x1, and x2, and the function values at those points, f0, f1, f2, compute the second-order approximation of the second derivative at x1

f(x) = x^4
x0, x1, x2 = 1.9999, 2.0, 2.0001
sd = second_deriv_central_diff(f(x0), f(x1), f(x2), x0, x1, x2)
sd ≈ 48

# output

true

HallThruster.second_deriv_coeffs — Method

second_deriv_coeffs(x0, x1, x2)

Generate finite difference coefficients for a central second derivative approximation at the point x1 on a three-point stencil at points x0, x1, and x2

julia> second_deriv_coeffs(-2//1, 0//1, 2//1)
(1//4, -1//2, 1//4)
julia> second_deriv_coeffs(-1//2, 0//1, 1//1)
(8//3, -4//1, 4//3)

HallThruster.sheath_potential — Method

sheath_potential(Tev, γ, mi))

compute wall sheath to be used for radiative losses and loss to wall. Goebel Katz equ. 7.3-29, 7.3-44. Assumed nₑuₑ/nᵢuᵢ ≈ 0.5 Sheath potentials are positive by convention in HallThruster.jl.

HallThruster.solve_potential_edge! — Method

solve_potential_edge!(; U::Matrix{Float64}, params::NamedTuple)

function to solve the potential equation derived from the generalized Ohm's law and employing charge conservation using quasineutrality. Second derivatives approximated with 2nd order central difference scheme, first derivatives with central difference. Centers of computational mesh placed on edges of fluid mesh, therefore edges correspond to boundaries for fluid. If required, interpolation is used to infer properties at mesh boundaries. Potential is a function of magnetic field, anomalous and classical collision frequency, neutral and ion density as well as ion velocity, and electron density and temperature leading to electron pressure. Solved by applying Thomas algorithm, which is of complexity O(n) and valid if matrix tridiagonal and diagonally dominant. The latter assumption almost always holds unless there is a huge discontinuity in either electron mobility or electron density.

HallThruster.supported_gases — Method

supported_gases(model::ReactionModel)::Vector{HallThruster.Gas}

Check which gases are supported by a given reaction model. If an empty vector is provided, then there are no restrictions on what gases can be used.

HallThruster.time_average — Function

time_average(sol, tstampstart)

compute time-averaged solution, input Solution type and the frame at which averaging starts. Returns a Solution object with a single frame.

HallThruster.upwind_diff_coeffs — Method

upwind_diff_coeffs(x0, x1, x2)

Generate finite difference coefficients for a upwind first derivative approximation at the point x2 on a three-point stencil at points x0, x1, and x2 (uses only points x0 and x1)

julia> upwind_diff_coeffs(-3//1, 0//1, 2//1)
(-1//3, 1//3, 0//1)

HallThruster.write_restart — Method

write_restart(path::AbstractString, sol)

Write a restart file to path`.

This can be reloaded to resume a simulation. The filetype can be anything supported by FileIO, though JLD2 is preferred.

HallThruster.σ_en — Method

σ_en(Tev)

Electron neutral collision cross section in m² as a function of electron temperature in eV. Eq. 3.6-13, from Fundamentals of Electric Propulsion, Goebel and Katz, 2008.